src/marketingkit-pro-backend/requirements.txt
flask
pillow
aura-sr